refactor(diff): per-provider modules + registry, drop github_ prefixes
Reorganise the PR-provider abstraction so adding a forge is one module plus
one registry entry, and provider-specific code stops leaking across files:

- pr_provider/ becomes a directory module: github.rs (all the gh-CLI logic,
  moved out of diff/mod.rs and git.rs and de-prefixed) and azure/ (mod.rs for
  URL/remote parsing + the PrProvider impl, client.rs for the REST client).
- Replace ProviderKind + handler() match + per-method dispatch with a
  &'static [&dyn PrProvider] registry; selection iterates it. Store the
  provider as &'static dyn PrProvider directly on PrInfo (trait gains : Sync
  so it can cross the viewed-sync threads), dropping the enum entirely.
- git.rs is now generic VCS only; mod.rs no longer holds github_* functions.

Behaviour-preserving; errors are still String and PrInfo still carries the
Azure Option fields (addressed in follow-up commits).
