+++ src/command/diff/pr_provider/azure/mod.rs
@@ -1,241 +1,15 @@
-//! Pull-request hosting provider abstraction.
-//!
-//! `lumen diff --pr` originally only understood GitHub (it shelled out to the
-//! `gh` CLI everywhere). This module introduces a [`PrProvider`] trait so other
-//! forges can be supported, with `gh` for GitHub and `az` (the Azure DevOps CLI)
-//! for Azure DevOps. New providers (e.g. `glab` for GitLab) only need to add an
-//! impl and wire it into provider detection.
+//! Azure DevOps provider: URL/remote parsing and the [`PrProvider`] impl. The
+//! REST client lives in [`client`].
 
-use std::collections::HashSet;
+mod client;
+
 use std::process::Command;
-use std::thread;
 
-use super::azure;
-use super::git::{github_load_pr_file_diffs, percent_encode};
-use super::types::FileDiff;
-use super::{
-    github_detect_current_branch_pr, github_fetch_pr_info, github_fetch_viewed_files,
-    github_mark_file_as_viewed_sync, github_unmark_file_as_viewed_sync, PrInfo,
-};
+use crate::command::diff::git::percent_encode;
+use crate::command::diff::types::FileDiff;
+use crate::command::diff::PrInfo;
 
-/// Which hosting provider a [`PrInfo`] belongs to. Stored on `PrInfo` so the
-/// runtime (viewed-file sync, open-in-browser, reloads) can route back to the
-/// right provider without re-parsing the original input.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ProviderKind {
-    GitHub,
-    Azure,
-}
-
-impl ProviderKind {
-    /// The provider implementation. Both providers are zero-sized, so this hands
-    /// back a `'static` reference with no allocation.
-    pub fn handler(self) -> &'static dyn PrProvider {
-        match self {
-            ProviderKind::GitHub => &GitHubProvider,
-            ProviderKind::Azure => &AzureProvider,
-        }
-    }
-}
-
-/// A pull-request hosting provider. Each method maps to one capability the diff
-/// UI needs; the viewed-file sync methods default to no-ops so providers without
-/// that concept (Azure DevOps) don't have to implement them.
-pub trait PrProvider {
-    /// Does this provider recognise `input` as one of its PR URLs?
-    fn matches_url(&self, input: &str) -> bool;
-
-    /// Does this provider recognise `origin` (a git remote URL) as one of its
-    /// repositories? Used to pick a provider for bare PR numbers.
-    fn matches_origin(&self, origin: &str) -> bool;
-
-    /// Resolve a PR number/URL into full metadata.
-    fn fetch_pr_info(&self, input: &str, repo_override: Option<&str>) -> Result<PrInfo, String>;
-
-    /// Find the PR associated with the current branch.
-    fn detect_current_branch_pr(&self, repo_override: Option<&str>) -> Result<String, String>;
-
-    /// Load the file diffs for a PR.
-    fn load_pr_file_diffs(&self, pr: &PrInfo) -> Result<Vec<FileDiff>, String>;
-
-    /// Whether this provider supports syncing per-file "viewed" state.
-    fn supports_viewed_sync(&self) -> bool {
-        false
-    }
-
-    /// Fetch the set of paths currently marked as viewed.
-    fn fetch_viewed_files(&self, _pr: &PrInfo) -> Result<HashSet<String>, String> {
-        Ok(HashSet::new())
-    }
-
-    /// Mark/unmark a file as viewed (blocking).
-    fn set_file_viewed(&self, _pr: &PrInfo, _path: &str, _viewed: bool) -> Result<(), String> {
-        Ok(())
-    }
-
-    /// Build a browser URL for `filename` within the PR.
-    fn file_web_url(&self, pr: &PrInfo, filename: &str) -> Option<String>;
-}
-
-// ---------------------------------------------------------------------------
-// Provider selection
-// ---------------------------------------------------------------------------
-
-/// True if `input` looks like a PR reference (a known PR URL or a bare number).
-pub fn is_pr_reference(input: &str) -> bool {
-    GitHubProvider.matches_url(input)
-        || AzureProvider.matches_url(input)
-        || input.parse::<u64>().is_ok()
-}
-
-fn read_origin_url() -> Option<String> {
-    let output = Command::new("git")
-        .args(["remote", "get-url", "origin"])
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
-    }
-    let url = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    if url.is_empty() {
-        None
-    } else {
-        Some(url)
-    }
-}
-
-/// Pick a provider from the git `origin` remote (and any `--origin` override),
-/// defaulting to GitHub when nothing matches.
-fn provider_for_origin(repo_override: Option<&str>) -> &'static dyn PrProvider {
-    let candidates = [repo_override.map(|s| s.to_string()), read_origin_url()];
-    for candidate in candidates.into_iter().flatten() {
-        if AzureProvider.matches_origin(&candidate) {
-            return ProviderKind::Azure.handler();
-        }
-    }
-    ProviderKind::GitHub.handler()
-}
-
-/// Pick a provider from a PR URL/number, falling back to origin detection.
-fn provider_for_input(input: &str, repo_override: Option<&str>) -> &'static dyn PrProvider {
-    if AzureProvider.matches_url(input) {
-        return ProviderKind::Azure.handler();
-    }
-    if GitHubProvider.matches_url(input) {
-        return ProviderKind::GitHub.handler();
-    }
-    provider_for_origin(repo_override)
-}
-
-// ---------------------------------------------------------------------------
-// Dispatchers used by the rest of the diff UI
-// ---------------------------------------------------------------------------
-
-pub fn fetch_pr_info(input: &str, repo_override: Option<&str>) -> Result<PrInfo, String> {
-    provider_for_input(input, repo_override).fetch_pr_info(input, repo_override)
-}
-
-pub fn detect_current_branch_pr(repo_override: Option<&str>) -> Result<String, String> {
-    provider_for_origin(repo_override).detect_current_branch_pr(repo_override)
-}
-
-pub fn load_pr_file_diffs(pr: &PrInfo) -> Result<Vec<FileDiff>, String> {
-    pr.provider.handler().load_pr_file_diffs(pr)
-}
-
-pub fn fetch_viewed_files(pr: &PrInfo) -> Result<HashSet<String>, String> {
-    pr.provider.handler().fetch_viewed_files(pr)
-}
-
-pub fn pr_file_web_url(pr: &PrInfo, filename: &str) -> Option<String> {
-    pr.provider.handler().file_web_url(pr, filename)
-}
-
-pub fn mark_file_as_viewed_async(pr: &PrInfo, file_path: &str) {
-    set_file_viewed_async(pr, file_path, true);
-}
-
-pub fn unmark_file_as_viewed_async(pr: &PrInfo, file_path: &str) {
-    set_file_viewed_async(pr, file_path, false);
-}
-
-fn set_file_viewed_async(pr: &PrInfo, file_path: &str, viewed: bool) {
-    if !pr.provider.handler().supports_viewed_sync() {
-        return;
-    }
-    let pr = pr.clone();
-    let path = file_path.to_string();
-    thread::spawn(move || {
-        let _ = pr.provider.handler().set_file_viewed(&pr, &path, viewed);
-    });
-}
-
-/// SHA-256 file anchor used by GitHub's PR "Files changed" deep links
-/// (`#diff-<sha256(path)>`).
-fn github_file_anchor(filename: &str) -> String {
-    use sha2::{Digest, Sha256};
-    let mut hasher = Sha256::new();
-    hasher.update(filename.as_bytes());
-    format!("{:x}", hasher.finalize())
-}
-
-// ---------------------------------------------------------------------------
-// GitHub
-// ---------------------------------------------------------------------------
-
-pub struct GitHubProvider;
-
-impl PrProvider for GitHubProvider {
-    fn matches_url(&self, input: &str) -> bool {
-        input.starts_with("http") && input.contains("/pull/")
-    }
-
-    fn matches_origin(&self, origin: &str) -> bool {
-        origin.contains("github.com")
-    }
-
-    fn fetch_pr_info(&self, input: &str, repo_override: Option<&str>) -> Result<PrInfo, String> {
-        github_fetch_pr_info(input, repo_override)
-    }
-
-    fn detect_current_branch_pr(&self, _repo_override: Option<&str>) -> Result<String, String> {
-        github_detect_current_branch_pr()
-    }
-
-    fn load_pr_file_diffs(&self, pr: &PrInfo) -> Result<Vec<FileDiff>, String> {
-        github_load_pr_file_diffs(pr)
-    }
-
-    fn supports_viewed_sync(&self) -> bool {
-        true
-    }
-
-    fn fetch_viewed_files(&self, pr: &PrInfo) -> Result<HashSet<String>, String> {
-        github_fetch_viewed_files(pr)
-    }
-
-    fn set_file_viewed(&self, pr: &PrInfo, path: &str, viewed: bool) -> Result<(), String> {
-        if viewed {
-            github_mark_file_as_viewed_sync(&pr.node_id, path)
-        } else {
-            github_unmark_file_as_viewed_sync(&pr.node_id, path)
-        }
-    }
-
-    fn file_web_url(&self, pr: &PrInfo, filename: &str) -> Option<String> {
-        Some(format!(
-            "https://github.com/{}/{}/pull/{}/files#diff-{}",
-            pr.repo_owner,
-            pr.repo_name,
-            pr.number,
-            github_file_anchor(filename)
-        ))
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Azure DevOps
-// ---------------------------------------------------------------------------
+use super::{read_origin_url, PrProvider};
 
 pub struct AzureProvider;
 
@@ -293,10 +67,10 @@ impl PrProvider for AzureProvider {
             .id
             .ok_or_else(|| format!("No PR id found in: {}", input))?;
 
-        let meta = azure::fetch_pr_metadata(&az.org_url, &az.project, &az.repo, id)?;
+        let meta = client::fetch_pr_metadata(&az.org_url, &az.project, &az.repo, id)?;
 
         Ok(PrInfo {
-            provider: ProviderKind::Azure,
+            provider: &AzureProvider,
             number: id,
             node_id: String::new(),
             repo_owner: az.org.clone(),
@@ -333,12 +107,12 @@ impl PrProvider for AzureProvider {
             return Err("Could not determine the current branch".to_string());
         }
 
-        let id = azure::detect_active_pr(&az.org_url, &az.project, &az.repo, &branch)?;
+        let id = client::detect_active_pr(&az.org_url, &az.project, &az.repo, &branch)?;
         Ok(id.to_string())
     }
 
     fn load_pr_file_diffs(&self, pr: &PrInfo) -> Result<Vec<FileDiff>, String> {
-        azure::load_pr_file_diffs(pr)
+        client::load_pr_file_diffs(pr)
     }
 
     fn file_web_url(&self, pr: &PrInfo, filename: &str) -> Option<String> {
@@ -475,13 +249,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn github_matches_pull_urls() {
-        assert!(GitHubProvider.matches_url("https://github.com/owner/repo/pull/123"));
-        assert!(!GitHubProvider.matches_url("https://dev.azure.com/o/p/_git/r/pullrequest/1"));
-        assert!(GitHubProvider.matches_origin("git@github.com:owner/repo.git"));
-    }
-
-    #[test]
     fn azure_matches_pr_urls() {
         assert!(AzureProvider.matches_url("https://dev.azure.com/o/p/_git/r/pullrequest/42"));
         assert!(AzureProvider.matches_url("https://myorg.visualstudio.com/p/_git/r/pullrequest/7"));
@@ -543,13 +310,5 @@ mod tests {
     #[test]
     fn encodes_path_query() {
         assert_eq!(percent_encode("/src/main.rs"), "%2Fsrc%2Fmain.rs");
-    }
-
-    #[test]
-    fn is_pr_reference_detects_forms() {
-        assert!(is_pr_reference("123"));
-        assert!(is_pr_reference("https://github.com/o/r/pull/1"));
-        assert!(is_pr_reference("https://dev.azure.com/o/p/_git/r/pullrequest/1"));
-        assert!(!is_pr_reference("main..feature"));
     }
 }
